+++ CHANGELOG.md
@@ -13,8 +13,12 @@
 
 ### [1.6.0]
 
+#### Added
+* `SleepWithTitle(int duration)` method added instead of `Sleep(int duration)`. It writes sleeping duration into Console's title.
+
 #### Changed
 * `Beep(Melody melody)` was checking `Melody` elements with checking duration of beeping and duration of waiting as not being lower than zero. Now it is checking with equal or lower than zero as it should be.
+* `Sleep(int duration)` now only waits instead of writing sleeping duration into screen.
 
 ### [1.5.0]
 #### Added